airflow/requirements.txt
gssapi==1.7.3
apache-airflow==2.2.5
kubernetes==11.0.0
pyspark==3.2.1
# snakebite-py3==3.0.5
apache-airflow-providers-apache-spark==2.1.3
apache-airflow-providers-ssh==2.4.3
apache-airflow-providers-telegram==2.0.4
apache-airflow-providers-cncf-kubernetes==3.0.0
apache-airflow-providers-apache-hive==2.3.2
# apache-airflow-providers-apache-hdfs==2.2.3
kafka-python==2.0.2
airflow-provider-kafka==0.1.0